+++ src/devices.c
@@ -24,14 +24,21 @@
 
 #include "6809cpu.h"
 #include "to8demulator.h"
+#include "sap.h"
+
+#define SECTOR_SIZE      256  // Size in bytes of a double density sector
+#define NB_TRACKS         80  // Number of tracks in a floppy
+#define SECTORS_PER_TRACK 16  // Number of sectors in a track
+#define SECTORS_PER_SIDE  SECTORS_PER_TRACK * NB_TRACKS
 
 // Global variables
 static bool fdprotection = true;
 static bool k7protection = true;
 static bool printerEnabled = false;
-static FILE *ffd = NULL;   // floppy file
+static FILE *ffd = NULL;   // floppy file (fd format)
 static FILE *fk7 = NULL;   // tape file
 static FILE *fprn = NULL;  // printer file
+static SapFile sap = { 0, NULL }; // floppy file (sap format)
 
 // 6809 registers
 #define CC dc6809_cc
@@ -73,88 +80,125 @@ static void Diskerror(int n)
   return;
 }
 
-// Floppy drive: Read a sector
+// Floppy drive: Read a sector.
+// This function emulates the DKCO function of the monitor with DK.OPC=2.
 static void Readsector(void)
 {
-  char buffer[256];
+  char buffer[SECTOR_SIZE];
   int i, j, u, p, s;
-  if(ffd == NULL) {Diskerror(71); return;}
-  u = Mgetc(0x6049) & 0xff; if(u > 03) {Diskerror(53); return;}
+  if (ffd == NULL && sap.handle == NULL) {Diskerror(71); return;}
+  // Drive number (0/1: 2 sides of the internal drive,
+  //               2/3: 2 sides of the external drive,
+  //               4  : RAM disk)
+  u = Mgetc(0x6049) & 0xff; if(u > 3) {Diskerror(53); return;}
+  // Track number (0->79)
   p = Mgetc(0x604a) & 0xff; if(p != 0) {Diskerror(53); return;}
-  p = Mgetc(0x604b) & 0xff; if(p > 79) {Diskerror(53); return;}
-  s = Mgetc(0x604c) & 0xff; if((s == 0) || (s > 16)) {Diskerror(53); return;}
-  s += 16 * p + 1280 * u;
-  if (fseek(ffd, 0, SEEK_END)) {Diskerror(53); return;}
-  if((s << 8) > ftell(ffd)) {Diskerror(53); return;}
-  for(j = 0; j < 256; j++) buffer[j] = 0xe5;
-  if (fseek(ffd, (s - 1) << 8, SEEK_SET)) {Diskerror(53); return;}
+  p = Mgetc(0x604b) & 0xff; if(p >= NB_TRACKS) {Diskerror(53); return;}
+  // Sector number
+  s = Mgetc(0x604c) & 0xff; if((s == 0) || (s > SECTORS_PER_TRACK)) {Diskerror(53); return;}
+  for (j = 0; j < SECTOR_SIZE; j++) buffer[j] = 0xe5;
+  if (ffd != NULL)
+  {
+    // FD file
+    s += SECTORS_PER_TRACK * p + SECTORS_PER_SIDE * u;
+    if (fseek(ffd, 0, SEEK_END)) {Diskerror(53); return;}
+    if ((s << 8) > ftell(ffd)) {Diskerror(53); return;}
+    if (fseek(ffd, (s - 1) << 8, SEEK_SET)) {Diskerror(53); return;}
+    if (fread(buffer, SECTOR_SIZE, 1, ffd) == 0) {Diskerror(53); return;}
+  }
+  else
+  {
+    // SAP file
+    if (!sap_readSector(&sap, p, s, buffer)) {Diskerror(53); return;}
+  }
   i = ((Mgetc(0x604f) & 0xff) << 8) + (Mgetc(0x6050) & 0xff);
-  if(fread(buffer, 256, 1, ffd) == 0) {Diskerror(53); return;}
-  for(j = 0; j < 256; j++) Mputc(i++, buffer[j]);
+  for (j = 0; j < SECTOR_SIZE; j++) Mputc(i++, buffer[j]);
 }
 
-// Floppy drive: Write a sector
+// Floppy drive: Write a sector.
+// This function emulates the DKCO function of the monitor with DK.OPC=8.
 static void Writesector(void)
 {
-  char buffer[256];
+  char buffer[SECTOR_SIZE];
   int i, j, u, p, s;
-  if(ffd == NULL) {Diskerror(71); return;}
+  if (ffd == NULL && sap.handle == NULL) {Diskerror(71); return;}
   // error 72 = write protection
-  if(fdprotection) {Diskerror(72); return;}
-  u = Mgetc(0x6049) & 0xff; if(u > 03) {Diskerror(53); return;}
+  if (fdprotection) {Diskerror(72); return;}
+  // Drive number (0/1: 2 sides of the internal drive,
+  //               2/3: 2 sides of the external drive,
+  //               4  : RAM disk)
+  u = Mgetc(0x6049) & 0xff; if(u > 3) {Diskerror(53); return;}
+  // Track number (0->79)
   p = Mgetc(0x604a) & 0xff; if(p != 0) {Diskerror(53); return;}
-  p = Mgetc(0x604b) & 0xff; if(p > 79) {Diskerror(53); return;}
-  s = Mgetc(0x604c) & 0xff; if((s == 0) || (s > 16)) {Diskerror(53); return;}
-  s += 16 * p + 1280 * u;
-  if (fseek(ffd, (s - 1) << 8, SEEK_SET)) {Diskerror(53); return;}
-  i = 256 * (Mgetc(0x604f) & 0xff) + (Mgetc(0x6050) & 0xff);
-  for(j = 0; j < 256; j++) buffer[j] = Mgetc(i++);
-  if(fwrite(buffer, 256, 1, ffd) == 0) Diskerror(53);
+  p = Mgetc(0x604b) & 0xff; if(p >= NB_TRACKS) {Diskerror(53); return;}
+  // Sector number
+  s = Mgetc(0x604c) & 0xff; if((s == 0) || (s > SECTORS_PER_TRACK)) {Diskerror(53); return;}
+  i = SECTOR_SIZE * (Mgetc(0x604f) & 0xff) + (Mgetc(0x6050) & 0xff);
+  for (j = 0; j < SECTOR_SIZE; j++) buffer[j] = Mgetc(i++);
+  if (ffd != NULL)
+  {
+    // FD file
+    s += SECTORS_PER_TRACK * p + SECTORS_PER_SIDE * u;
+    if (fseek(ffd, (s - 1) << 8, SEEK_SET)) {Diskerror(53); return;}
+    if (fwrite(buffer, SECTOR_SIZE, 1, ffd) == 0) {Diskerror(53); return;}
+  }
+  else
+  {
+    // SAP file
+    if (!sap_writeSector(&sap, p, s, buffer)) {Diskerror(53); return;}
+  }
 }
 
-// Format a floppy disk
+// Format a floppy disk.
+// This function emulates the DKFORM function of the monitor.
 static void Formatdisk(void)
 {
-  char buffer[256];
+  char buffer[SECTOR_SIZE];
   int i, u, fatlength;
-  if(ffd == NULL) {Diskerror(71); return;}
+  if (ffd == NULL) {Diskerror(71); return;}
   // error 72 = write protection
-  if(fdprotection) {Diskerror(72); return;}
+  if (fdprotection) {Diskerror(72); return;}
   u = Mgetc(0x6049) & 0xff; if(u > 03) return; // Unit
-  u = (1280 * u) << 8; // Start of the unit in the .fd file
+  u = (SECTORS_PER_SIDE * u) << 8; // Start of the unit in the .fd file
   fatlength = 160;     // 80=160Ko, 160=320Ko
-  //rem: fatlength provisoire !!!!! (tester la variable adequate)
+  // rem: fatlength provisoire !!!!! (tester la variable adequate)
   // Initialisation of the whole disk with 0xE5
-  for(i = 0; i < 256; i++) buffer[i] = 0xe5;
+  for (i = 0; i < SECTOR_SIZE; i++) buffer[i] = 0xe5;
   if (fseek(ffd, u, SEEK_SET)) {Diskerror(53); return;}
-  for(i = 0; i < (fatlength * 8); i++)
-    if(fwrite(buffer, 256, 1, ffd) == 0) {Diskerror(53); return;}
+  for (i = 0; i < (fatlength * 8); i++)
+    if (fwrite(buffer, SECTOR_SIZE, 1, ffd) == 0) {Diskerror(53); return;}
   // Initialisation of track 20 at 0xFF
-  for(i = 0; i < 256; i++) buffer[i] = 0xff;
+  for (i = 0; i < SECTOR_SIZE; i++) buffer[i] = 0xff;
   if (fseek(ffd, u + 0x14000, SEEK_SET)) {Diskerror(53); return;}
-  for(i = 0; i < 16; i++)
-    if(fwrite(buffer, 256, 1, ffd) == 0) {Diskerror(53); return;}
+  for (i = 0; i < SECTORS_PER_TRACK; i++)
+    if (fwrite(buffer, SECTOR_SIZE, 1, ffd) == 0) {Diskerror(53); return;}
   // Write the FAT
   buffer[0x00] = 0;
   buffer[0x29] = 0xfe; buffer[0x2a] = 0xfe;
-  for(i = fatlength + 1; i < 256; i++) buffer[i] = 0xfe;
+  for (i = fatlength + 1; i < SECTOR_SIZE; i++) buffer[i] = 0xfe;
   if (fseek(ffd, u + 0x14100, SEEK_SET)) {Diskerror(53); return;}
-  if(fwrite(buffer, 256, 1, ffd) == 0) {Diskerror(53); return;}
+  if (fwrite(buffer, SECTOR_SIZE, 1, ffd) == 0) {Diskerror(53); return;}
 }
 
-void Unloadfd(void)
+void UnloadFloppy(void)
 {
-  if(ffd) {fclose(ffd); ffd = NULL;}
+  if (ffd) {fclose(ffd); ffd = NULL;}
+  if (sap.handle) {sap_close(&sap);}
 }
 
-void Loadfd(const char *filename)
+void LoadFd(const char *filename)
 {
   //fermeture disquette eventuellement ouverte
-  Unloadfd();
+  UnloadFloppy();
   if(filename[0] == '\0') return;
   //ouverture de la nouvelle disquette
   ffd = fopen(filename, "rb+");
-  if(ffd == NULL) return;
+}
+
+void LoadSap(const char *filename)
+{
+  UnloadFloppy();
+  sap = sap_open(filename);
 }
 
 // Tape drive: read a byte
@@ -181,14 +225,14 @@ static void Writebytek7(void)
   Mputc(0x2045, 0);
 }
 
-void Unloadk7(void)
+void UnloadK7(void)
 {
   if(fk7) {fclose(fk7); fk7 = NULL;}
 }
 
-void Loadk7(const char *filename)
+void LoadK7(const char *filename)
 {
-  Unloadk7();
+  UnloadK7();
   if(filename[0] == '\0') return;
   fk7 = fopen(filename, "rb+");
   if(fk7 == NULL) return;
@@ -196,19 +240,19 @@ void Loadk7(const char *filename)
   fseek(fk7, 0, SEEK_SET);
 }
 
-void Unloadmemo(void)
+void UnloadMemo(void)
 {
   carflags = 0;
   Hardreset();
 }
 
-void Loadmemo(const char *filename)
+void LoadMemo(const char *filename)
 {
   FILE *fp = NULL;
   int i, c, carsize;
   // Open the memo7 file
   fp = fopen(filename, "rb");
-  if(fp == NULL) {Unloadmemo(); return;}
+  if(fp == NULL) {UnloadMemo(); return;}
   // Loading
   carsize = 0;
   while(((c = fgetc(fp)) != EOF) && (carsize < 0x10000)) car[carsize++] = c;